Split workspace chrome Playwright flows
